Day 9 | IntCodeComputer Memory Supports Memory Address > Initial Snapshot. Fixed 64bit Int Support

+++ src/main/kotlin/com/aoc/intcode/instructions/InstructionStrategy.kt
@@ -9,16 +9,16 @@ interface InstructionStrategy {
 
     fun getValue(memorySnapshot: Memory, mode: ParameterMode, addressesAfterPointer: Int): Long {
         val parameterIndex = memorySnapshot.instructionPointer + addressesAfterPointer
-        val valueIndex = memorySnapshot.getInstructionAtAddress(parameterIndex).toInt()
+        val valueIndex = memorySnapshot.getInstructionAtAddress(parameterIndex)
         return when (mode) {
             ParameterMode.POSITION -> memorySnapshot.getInstructionAtAddress(valueIndex)
-            ParameterMode.IMMEDIATE -> valueIndex.toLong()
+            ParameterMode.IMMEDIATE -> valueIndex
             ParameterMode.RELATIVE -> memorySnapshot.getInstructionAtAddress(memorySnapshot.relativeBase + valueIndex)
         }
     }
 
-    fun getWriteToIndex(memorySnapshot: Memory, addressesAfterPointer: Int): Int {
+    fun getWriteToIndex(memorySnapshot: Memory, addressesAfterPointer: Int): Long {
         val parameterIndex = memorySnapshot.instructionPointer + addressesAfterPointer
-        return memorySnapshot.getInstructionAtAddress(parameterIndex).toInt()
+        return memorySnapshot.getInstructionAtAddress(parameterIndex)
     }
 }
